Started working on feature 6
Came to snag with wanting to both capture and display command output in
playlist. Likely going to need to create a pipe for both stdout and
stderr, spawn two threads: one to run the command, and write to the
pipes, and one to both capture and print out the bytes from the pipe.

Added the playlist.rs and rename.rs files, and linked to them
inside the main.rs file.

Added an eighth feature to create a verbose mode for each subcommand.